+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # domino x.x.x (not yet released)
 * Fix broken stopping of immediate propagation of Events. (#78)
 * Properly set "scripting enabled" flag when parsing fragments.
+* Fix handling of escaped or invalid CSS identifiers in
+  `querySelector` and friends. (#79)
 
 # domino 1.0.24 (05 Apr 2016)
 * Implement WindowTimers interface on Window. (#72)